**Funcionalidad:** Mejora de los tests de integración, unitarios y end to end de la prevención de compras sobre la capacidad del evento y la funcionalidad evaluación de los eventos; también se ha eliminado el modelo de usuario de terceros y se ha suplantado por el modelo ya creado previamente en nuestro proyecto.

## Incluye en la funcionalidad:
Tests antiguos:
- ratings e2e 26.160 s
- overselling e2e 10.575 s
- ratings integration 12.929 s
- ratings unit 6.903 s
- overselling unit 9.733 s

Tests nuevos:
- ratings e2e 23446 s
- overselling e2e 9.990 s
- ratings integration 12.139 s
- ratings unit 6.3 s
- overselling unit 3.5 s