@PietrantuonoFranco 2025-06-10
## Errores:
### Tests unitarios
En ambos test unitarios se repite que se crean varias instancias para testear distintas cosas (en un test creas 2 instancias, en el iguiente otras dos y así sucesivamente). Sería mejor que crees usuarios y tickets como base en el método `setUp()`, Esto te va a permitir no estar creando instancias constantemente en cada test y en su lugar crear solo la que necesites en el mismo.
_Por ejemplo:_
Podrías crear:
    - una venue con capacidad de 5,
    - dos usuarios,
    - un ticket,
    - etc

### Tests E2E
En cuanto a los test E2E, podrías implementar las instancias ya creadas en el `setUp()` del archivo `base.py` y en caso de ser necesario crear alguna instancia para tus test en específico, los crearías desde el `setUp()` en el archivo de tu test.
En la línea 72, aparace la siguiente línea `event_id = cast(int, self.event.id)`. Esta es innecesaria ya que podrías llamar directamente a la instancia en las distintas URL como `self.event.id`, sin necesidad de utilizar el método `cast()`.

> [!TIP]
> **Tene en cuenta lo siguiente:**
> - Los cambios en `setUp()` no se revierten después de cada test a diferencia que los realizados durante la ejecución del test, los cuales se revierten automáticamente.
> - Cada test empieza con un estado limpio (aislamiento garantizado).